--- hpp/irc.hpp
@@ -1,12 +1,0 @@
-#ifndef IRC_HPP
-#define IRC_HPP
-
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-#include <cstdlib>
-#include <iostream>
-#include <unistd.h>
-
-
-#endif
