+++ srcs/Client.class.cpp
@@ -1,27 +1,48 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.cpp                                           :+:      :+:    :+:   */
+/*   Client.class.cpp                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: soutin <soutin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/08/03 16:09:02 by soutin            #+#    #+#             */
-/*   Updated: 2024/08/03 20:00:57 by soutin           ###   ########.fr       */
+/*   Created: 2024/08/13 18:02:46 by soutin            #+#    #+#             */
+/*   Updated: 2024/08/13 18:02:47 by soutin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "Server.hpp"
+#include "Client.class.hpp"
 
-int	main(int ac, char **av)
+Client::Client()
 {
-	if (ac != 3)
-	{
-		std::cout << "Usage: ./ircserv [port] [password]\n";
-		return (1);
-	}
-	
-	Server	server(atoi(av[1]), av[2]);
-	server.startServer();
-	server.run();
-	return (0);
+  _pass = false;
+}
+
+Client::~Client() {}
+
+std::string Client::getNick()
+{
+  return _nick;
+}
+
+bool Client::checkPass()
+{
+  return _pass;
+}
+
+std::string Client::getUser()
+{
+  return _user;
+}
+
+void Client::setNick(std::string nick)
+{
+  _nick = nick;
+}
+void Client::setPass(bool pass)
+{
+  _pass = pass;
+}
+void Client::setUser(std::string user)
+{
+  _user = user;
 }
