+++ srcs/main.cpp
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   main.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: soutin <soutin@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/08/03 16:09:02 by soutin            #+#    #+#             */
+/*   Updated: 2024/08/13 18:02:35 by soutin           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ft_irc.hpp"
+
+int	main(int ac, char **av)
+{
+	if (ac != 3)
+	{
+		std::cout << "Usage: ./ircserv [port] [password]\n";
+		return (1);
+	}
+	
+	Server	server(atoi(av[1]), av[2]);
+	server.startServer();
+	server.run();
+	return (0);
+}
