+++ srcs/utils.cpp
@@ -1,0 +1,36 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.cpp                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: soutin <soutin@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/08/13 18:02:53 by soutin            #+#    #+#             */
+/*   Updated: 2024/08/13 18:02:54 by soutin           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "ft_irc.hpp"
+
+bool verify_string_format(const std::string &input_string)
+{
+	for (size_t i = 0; i < input_string.size(); i++)
+	{
+		char c = input_string[i];
+		if (!((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || (c >= '0' && c <= '9') || std::string("-[]\\`^{}").find(c) != std::string::npos))
+			return (false);
+	}
+	return (true);
+}
+
+bool nickAlreadyInUse(std::map<int, Client> *clients, std::string arg, int clientFd)
+{
+	std::map<int, Client>::iterator it;
+
+	for (it = clients->begin(); it != clients->end(); it++)
+	{
+		if (!it->second.getNick().compare(arg) && it->first != clientFd)
+			return (true);
+	}
+	return (false);
+}
